* Fixed file transfers on 32-bit systems (sometimes size_t matters). * Fixed some minor bugs on the go.

+++ transfer.h
@@ -63,7 +63,7 @@ struct TRANSFER_T_STRUCT {
 
 extern transfer_t *offer_new( uint64_t dest, const char *filename );
 extern transfer_t *offer_match( uint64_t oid, uint64_t rid );
-extern void *offer_read( transfer_t *o, uint64_t off, size_t *psz );
+extern void *offer_read( transfer_t *o, uint64_t off, uint64_t *psz );
 
 extern transfer_t *download_new( void );
 extern transfer_t *download_match( uint64_t oid );
